+++ alteia/core/resources/sharetokens.py
@@ -1,8 +1,0 @@
-from typing import List, NamedTuple
-
-from alteia.core.utils.typing import ShareToken
-
-ShareTokensWithTotal = NamedTuple(
-    'ShareTokensWithTotal',
-    [('total', int), ('results', List[ShareToken])]
-)